Make serial input work again
Instead of using signal to notify main thread to read more bytes, the
worker thread can read and fill the buffer by itself. If the rx_buf is
full, wait on the condition, the main thread will notify the worker
thread when the level of the rx_buf is below the half size.

Because two threads may modify the device registers and rx_buf, use a
mutex lock to protect them.